Revert "Separate mapreduce code from core (#425)"
* This reverts commit 67a1c696b7ab0415d1b216110d472c37d505f6a9.
* Return mapreduce code to core but maintain clean up for 2.0
