+++ pom.xml
@@ -78,7 +78,6 @@
   </mailingLists>
   <modules>
     <module>assemble</module>
-    <module>client/mapreduce</module>
     <module>core</module>
     <module>iterator-test-harness</module>
     <module>minicluster</module>
@@ -308,11 +307,6 @@
         <groupId>log4j</groupId>
         <artifactId>log4j</artifactId>
         <version>1.2.17</version>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.accumulo</groupId>
-        <artifactId>accumulo-client-mapreduce</artifactId>
-        <version>${project.version}</version>
       </dependency>
       <dependency>
         <groupId>org.apache.accumulo</groupId>
